Delete group functionality added

+++ app/src/main/java/com/example/raajesharunachalam/taskmanager/endpoints/GroupEndpoints.java
@@ -33,11 +33,11 @@ public interface GroupEndpoints {
 
     @Headers("Token: 5c8ab94e-3c95-40f9-863d-e31ae49e5d8d")
     @PUT("/groups/{gid}")
-    Call<Void> updateGroup(@Path("gid") int gid, @Body UpdateGroupRequest updateGroupRequest);
+    Call<Void> updateGroup(@Path("gid") long gid, @Body UpdateGroupRequest updateGroupRequest);
 
     @Headers("Token: 5c8ab94e-3c95-40f9-863d-e31ae49e5d8d")
     @DELETE("/groups/{gid}")
-    Call<Void> deleteGroup(@Path("gid") int gid);
+    Call<Void> deleteGroup(@Path("gid") long gid);
 
     Retrofit retrofit = new Retrofit.Builder()
             .baseUrl("http://taskmanager.host")
